hash в вишлисте

+++ apps/lk/wishlist/views.py
@@ -8,6 +8,7 @@ from django.http import Http404, JsonResponse
 from django.views.generic import View, TemplateView
 from django.utils import timezone
 
+from apps.cart.utils import make_hash_from_cartitem
 from apps.catalog.models import AttributeOption, Product, ProductOption
 from apps.core.mixins import JSONViewMixin
 from apps.lk.models import WishListItem
@@ -51,7 +52,8 @@ class WishListView(TemplateView):
             product_options = {o.id: o for o in ProductOption.objects.filter(id__in=product_options_ids)}
             options = {o.id: o for o in AttributeOption.objects.select_related('attribute').filter(id__in=options_ids)}
             wishlist = [item for item in wishlist
-                        if item['product_id'] in products.keys() and item.get('option_id') in product_options.keys()]
+                        if item['product_id'] in products.keys()
+                        and item.get('option_id') in product_options.keys()]
 
         self.wishlist = wishlist
         self.products = products
@@ -77,8 +79,10 @@ class WishListView(TemplateView):
                 'product_id': p.id,
                 'option_id': option.id,
                 'attrs': {k: v[0] for k, v in option.attrs.items()},
+                'extra_products': {},
                 'price': float(option.price),
             }
+            item['hash'] = make_hash_from_cartitem(item['attrs'], item['extra_products'])
             wishlist.append(item)
         self.request.session['wishlist'] = wishlist
         return wishlist
@@ -95,19 +99,24 @@ class WishListAddView(JSONViewMixin, View):
             option_id = int(DATA['option_id'])
             price = float(DATA.get('price', 0.0))
             attrs = DATA.get('attrs', {})
+            extra_products = DATA.get('extra_products', {})
         except ValueError:
             data = {'result': 'error', 'error': 'Неправильный формат запроса'}
             return JsonResponse(data, status=400)
 
+        hash = make_hash_from_cartitem(attrs, extra_products)
         wishlist = get_wishlist_from_request(self.request)
         # product_ids = [item['product_id'] for item in wishlist]
         # product_options_ids = [item['option_id'] for item in wishlist]
+
+        # print hash
 
         if profile.is_anonymous():
             the_item = None
 
             for item in wishlist:
-                if item['product_id'] == product_id:
+                # if item['product_id'] == product_id:
+                if item['product_id'] == product_id and item['hash'] == hash:
                     the_item = item
 
             if the_item is None:
@@ -118,20 +127,22 @@ class WishListAddView(JSONViewMixin, View):
                 'product_id': product_id,
                 'option_id': option_id,
                 'price': price,
-                'attrs': attrs
+                'attrs': attrs,
+                'extra_products': extra_products,
+                'hash': hash,
             })
             request.session['wishlist'] = wishlist
 
         else:
             try:
-                the_item, _created = WishListItem.objects.get_or_create(
-                    profile_id=profile.id,
-                    product_id=product_id,
-                    defaults={'option_id': option_id},
-                )
+                kwargs = {'profile_id': profile.id, 'product_id': product_id, 'hash': hash}
+                the_item = WishListItem.objects.filter(**kwargs).first()
+                if not the_item:
+                    the_item = WishListItem(**kwargs)
                 the_item.option_id = option_id
                 the_item.price = price
                 the_item.attrs = attrs
+                the_item.extra_products = extra_products
                 the_item.save()
             except IntegrityError:
                 data = {'result': 'error', 'error': 'Неправильный id товара'}
@@ -150,27 +161,29 @@ class WishListRemoveView(JSONViewMixin, View):
 
         try:
             product_id = int(DATA['product_id'])
+            attrs = DATA.get('attrs', {})
+            extra_products = DATA.get('extra_products', {})
         except ValueError:
             data = {'result': 'error', 'error': 'Неправильный формат запроса'}
             return JsonResponse(data, status=400)
 
+        hash = make_hash_from_cartitem(attrs, extra_products)
         wishlist = get_wishlist_from_request(self.request)
         product_ids = [item['product_id'] for item in wishlist]
 
         if product_id in product_ids:
             if profile.is_anonymous():
                 for i, item in enumerate(wishlist):
-                    if item['product_id'] == product_id:
+                    if item['product_id'] == product_id and item['hash'] == hash:
                         wishlist.pop(i)
                         break
                 request.session['wishlist'] = wishlist
 
             else:
                 try:
-                    the_item = WishListItem.objects.get(profile_id=profile.id, product_id=product_id)
-                    the_item.delete()
-                except WishListView.DoesNotExist:
-                    pass
+                    the_item = WishListItem.objects.filter(profile_id=profile.id, product_id=product_id, hash=hash).first()
+                    if the_item:
+                        the_item.delete()
                 except IntegrityError:
                     data = {'result': 'error', 'error': 'Неправильный id товара'}
                     return JsonResponse(data, status=400)
